+++ lib/mongoid/relations/builders/referenced/one.rb
@@ -17,7 +17,7 @@ module Mongoid # :nodoc:
           def build(type = nil)
             return object unless query?
             return nil if base.new_record?
-            metadata.criteria(Conversions.flag(object, metadata)).from_map_or_db
+            metadata.criteria(Conversions.flag(object, metadata), base.class).from_map_or_db
           end
         end
       end